BUG: bsplines fixed issue 14525
bspline, quadratic and cubic gave zero outputs for integer dtypes.
This was caused by initialization of the output to the input dtype in cubic, quadratic, and np.piecewise (in bspline). In cubic and quadratic this was modified in place. For bspline it was fixed by casting the input to float (after abs to avoid issues with complex input).

+++ scipy/signal/bsplines.py
@@ -186,7 +186,7 @@ def bspline(x, n):
     # number of pieces on the left-side is (n+1)/2
     funclist, condfuncs = _bspline_piecefunctions(n)
     condlist = [func(ax) for func in condfuncs]
-    return piecewise(ax, condlist, funclist)
+    return piecewise(ax.astype('float'), condlist, funclist)
 
 
 def gauss_spline(x, n):
@@ -281,7 +281,7 @@ def cubic(x):
 
     """
     ax = abs(asarray(x))
-    res = zeros_like(ax)
+    res = zeros_like(ax, dtype='float')
     cond1 = less(ax, 1)
     if cond1.any():
         ax1 = ax[cond1]
@@ -333,7 +333,7 @@ def quadratic(x):
 
     """
     ax = abs(asarray(x))
-    res = zeros_like(ax)
+    res = zeros_like(ax, dtype='float')
     cond1 = less(ax, 0.5)
     if cond1.any():
         ax1 = ax[cond1]
